move widgets and docs nav queries to component files
