+++ website/src/components/docs-nav.js
@@ -1,10 +1,16 @@
 import React, { Component } from 'react';
-import Link from 'gatsby-link';
+import { Link, StaticQuery, graphql } from 'gatsby';
 import classnames from 'classnames';
 
 import TableOfContents from './toc';
 
 import '../css/imports/docs-nav.css';
+
+const toMenu = (menu, nav) =>
+  menu.map(group => ({
+    title: group.title,
+    group: nav.group.find(g => g.fieldValue === group.name),
+  }));
 
 class DocsNav extends Component {
   state = {
@@ -18,49 +24,95 @@ class DocsNav extends Component {
   };
 
   render() {
-    const { items, location } = this.props;
+    const { location } = this.props;
     const { isOpen } = this.state;
 
     return (
-      <div className="docs-nav">
-        <button className="btn-primary docs-nav-btn" onClick={this.toggleNav}>
-          {isOpen ? <span>&times;</span> : <span>&#9776;</span>} {isOpen ? 'Hide' : 'Show'}{' '}
-          Navigation
-        </button>
-        <nav
-          className={classnames('docs-nav-content', {
-            'is-open': isOpen,
-          })}
-          id="docs-nav"
-        >
-          <ul className="docs-nav-list">
-            {items.map(item => (
-              <li className="docs-nav-item" key={item.title}>
-                <div className="docs-nav-title">{item.title}</div>
+      <StaticQuery
+        query={graphql`
+          query docsNavQuery {
+            nav: allMarkdownRemark(
+              sort: { fields: [frontmatter___weight], order: ASC }
+              filter: {
+                frontmatter: { title: { ne: null }, group: { ne: null } }
+                fields: { slug: { regex: "/docs/" } }
+              }
+            ) {
+              group(field: frontmatter___group) {
+                fieldValue
+                edges {
+                  node {
+                    fields {
+                      slug
+                    }
+                    frontmatter {
+                      title
+                      group
+                    }
+                    tableOfContents
+                  }
+                }
+              }
+            }
+            menu: site {
+              siteMetadata {
+                menu {
+                  docs {
+                    name
+                    title
+                  }
+                }
+              }
+            }
+          }
+        `}
+      >
+        {data => {
+          const docsNav = toMenu(data.menu.siteMetadata.menu.docs, data.nav);
+
+          return (
+            <div className="docs-nav">
+              <button className="btn-primary docs-nav-btn" onClick={this.toggleNav}>
+                {isOpen ? <span>&times;</span> : <span>&#9776;</span>} {isOpen ? 'Hide' : 'Show'}{' '}
+                Navigation
+              </button>
+              <nav
+                className={classnames('docs-nav-content', {
+                  'is-open': isOpen,
+                })}
+                id="docs-nav"
+              >
                 <ul className="docs-nav-list">
-                  {item.group.edges.map(({ node }) => (
-                    <li className="docs-nav-item" key={node.fields.slug}>
-                      <Link
-                        to={node.fields.slug}
-                        className="docs-nav-link"
-                        activeClassName="active"
-                        onClick={this.toggleNav}
-                      >
-                        {node.frontmatter.title}
-                      </Link>
-                      <div className="docs-nav-toc">
-                        {location.pathname === node.fields.slug && (
-                          <TableOfContents selector=".docs-content h2" />
-                        )}
-                      </div>
+                  {docsNav.map(item => (
+                    <li className="docs-nav-item" key={item.title}>
+                      <div className="docs-nav-title">{item.title}</div>
+                      <ul className="docs-nav-list">
+                        {item.group.edges.map(({ node }) => (
+                          <li className="docs-nav-item" key={node.fields.slug}>
+                            <Link
+                              to={node.fields.slug}
+                              className="docs-nav-link"
+                              activeClassName="active"
+                              onClick={this.toggleNav}
+                            >
+                              {node.frontmatter.title}
+                            </Link>
+                            <div className="docs-nav-toc">
+                              {location.pathname === node.fields.slug && (
+                                <TableOfContents selector=".docs-content h2" />
+                              )}
+                            </div>
+                          </li>
+                        ))}
+                      </ul>
                     </li>
                   ))}
                 </ul>
-              </li>
-            ))}
-          </ul>
-        </nav>
-      </div>
+              </nav>
+            </div>
+          );
+        }}
+      </StaticQuery>
     );
   }
 }
